requirements.txt
fastapi
uvicorn
pydantic
httpx
pluggy
PyYAML
xmlschema
python-dateutil
tenacity
